0.7.16
a final backup for the 0.7.x versions.
the syntax tree structure is backuped written by c and lex/yacc parser generator. we decide to go back to c-sharp platform if possible and then try the following improvements in scripting engine architecture.

  1. pointers allocated to objects once they are initialized, and base on these to create a pointer-driven reference and refcounting system.
  2. security in consideration, loops, iterations, callbacks and callstack depths in sandbox.
  3. full error messages in parser and tokenizer.
  4. strong type symbols, and data type operations.

these updates are scheduled in the 0.8.x versions. and the first commit after removal of old components will be published soon as 0.8.0, turing the 0.7.x engine to 'Simula.Scripting.Obsolete`